+++ backend/routes/index.js
@@ -1,7 +1,7 @@
 const router = require('express').Router()
 
-const drop = require('./drop/index');
-const user = require('./user/index'); 
+const drop = require('dropRouter');
+const user = require('userRouter'); 
 
 router.use('/drops', drop);
 router.use('/users', user);
